requirements.txt
langgraph>=0.2.30
langchain>=0.2.12
langchain-google-genai>=2.0.0
tavily-python>=0.4.0
python-dotenv>=1.0.1

